Add and check tokens in redis blacklist  (auth), encapsulate websocket and tkn functionalities
